refactor: remove deprecated ckb_js_tests
They're moved to tests/ts
