+++ package.json
@@ -10,7 +10,7 @@
     "format": "pnpm -r run format",
     "check:format": "pnpm -r run check:format",
     "publish": "pnpm build && pnpm -r publish",
-    "test": "pnpm --filter ckb-testtool --filter ckb-fs-packer --filter @ckb-js-std/examples --filter tests/ts run test",
+    "test": "pnpm --filter ckb-testtool --filter ckb-fs-packer --filter @ckb-js-std/examples --filter ts-tests run test",
     "docs": "typedoc"
   },
   "devDependencies": {
