--- tests/ckb_js_tests/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "ckb_js_tests"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ckb-debugger-api = { git = "https://github.com/nervosnetwork/ckb-standalone-debugger.git", tag="v0.111.0" }
-ckb-mock-tx-types =  { git = "https://github.com/nervosnetwork/ckb-standalone-debugger.git", tag="v0.111.0" }
-anyhow = "1.0.75"
-serde_json = "1.0.107"
