--- tests/ckb_js_tests/src/bin/load_script.rs
@@ -1,9 +1,0 @@
-use ckb_js_tests::read_tx_template;
-
-pub fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let tx = read_tx_template("templates/load_script.json")?;
-
-    let json = serde_json::to_string_pretty(&tx).unwrap();
-    println!("{}", json);
-    Ok(())
-}
