--- tests/ckb_js_tests/templates/simple_udt.json
@@ -1,100 +1,0 @@
-{
-  "mock_info": {
-    "inputs": [
-      {
-        "output": {
-          "capacity": "0x10000000",
-          "lock": {
-            "args": "0x",
-            "code_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-            "hash_type": "data1"
-          },
-          "type": {
-            "args": "0x0000{{ ref_type js-code-file }}01",
-            "code_hash": "0x{{ ref_type ckb-js-vm }}",
-            "hash_type": "type"
-          }
-        },
-        "data": "0x11000000000000000000000000000000"
-      }
-    ],
-    "cell_deps": [
-      {
-        "output": {
-          "capacity": "0x10000000",
-          "lock": {
-              "args": "0x00AE9DF3447C404A645BC48BEA4B7643B95AC5C3AE",
-              "code_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-              "hash_type": "data1"
-          },
-          "type": "{{ def_type ckb-js-vm }}"
-        },
-        "data": "0x{{ data ../../../build/ckb-js-vm }}"
-      },
-      {
-        "output": {
-          "capacity": "0x10000000",
-          "lock": {
-              "args": "0x00AE9DF3447C404A645BC48BEA4B7643B95AC5C3AE",
-              "code_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-              "hash_type": "data1"
-          },
-          "type": "{{ def_type js-code-file }}"
-        },
-        "data": "0x{{ data ../../../packages/examples/dist/simple_udt.bc }}"
-      }
-    ],
-    "header_deps": []
-  },
-  "tx": {
-    "outputs": [
-      {
-        "capacity": "0x0",
-        "lock": {
-          "args": "0x",
-          "code_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-          "hash_type": "type"
-        },
-        "type": {
-          "args": "0x0000{{ ref_type js-code-file }}01",
-          "code_hash": "0x{{ ref_type ckb-js-vm }}",
-          "hash_type": "type"
-        }
-      },
-      {
-        "capacity": "0x0",
-        "lock": {
-          "args": "0x",
-          "code_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-          "hash_type": "type"
-        },
-        "type": {
-          "args": "0x0000{{ ref_type js-code-file }}01",
-          "code_hash": "0x{{ ref_type ckb-js-vm }}",
-          "hash_type": "type"
-        }
-      },
-      {
-        "capacity": "0x0",
-        "lock": {
-          "args": "0x",
-          "code_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-          "hash_type": "type"
-        },
-        "type": {
-          "args": "0x0000{{ ref_type js-code-file }}01",
-          "code_hash": "0x{{ ref_type ckb-js-vm }}",
-          "hash_type": "type"
-        }
-      }
-    ],
-    "witnesses": [
-      "0x0001020304050607"
-    ],
-    "outputs_data": [
-      "0x10000000000000000000000000000000",
-      "0x01000000000000000000000000000000",
-      "0x00000000000000000000000000000000"
-    ]
-  }
-}
