--- tests/ckb_js_tests/test_data/fs_module_mount/fib_module.js
@@ -1,10 +1,0 @@
-/* fib module */
-export function fib(n)
-{
-    if (n <= 0)
-        return 0;
-    else if (n == 1)
-        return 1;
-    else
-        return fib(n - 1) + fib(n - 2);
-}
