--- tests/ckb_js_tests/test_data/load_script/fib.js
@@ -1,9 +1,0 @@
-function fib(n)
-{
-    if (n <= 0)
-        return 0;
-    else if (n == 1)
-        return 1;
-    else
-        return fib(n - 1) + fib(n - 2);
-}
