+++ tests/ts/jest.config.cjs
@@ -12,4 +12,5 @@ module.exports = {
       },
     ],
   },
+  maxWorkers: 1
 };
