+++ tests/ts/package.json
@@ -15,6 +15,7 @@
   "main": "./dist/index.js",
   "module": "./dist/index.js",
   "scripts": {
+    "build": "echo 'no build steps'",
     "clean": "rimraf dist",
     "format": "prettier --write .",
     "test": "jest"
