+++ tests/ts/src/file-system/data/init.js
@@ -1,7 +1,7 @@
 import * as ckb from "@ckb-js-std/bindings";
 
-ckb.mount(2, ckb.SOURCE_CELL_DEP, "/")
-ckb.mount(2, ckb.SOURCE_CELL_DEP, "/a")
-ckb.mount(2, ckb.SOURCE_CELL_DEP, "/b/")
+ckb.mount(2, ckb.SOURCE_CELL_DEP, "/");
+ckb.mount(2, ckb.SOURCE_CELL_DEP, "/a");
+ckb.mount(2, ckb.SOURCE_CELL_DEP, "/b/");
 
 console.log("init.js");
