+++ tests/ts/src/file-system/data2/fib.js
@@ -1,5 +1,4 @@
-/* fib module */
-export function fib(n) {
+function fib(n) {
   if (n <= 0) return 0;
   else if (n == 1) return 1;
   else return fib(n - 1) + fib(n - 2);
