--- tests/ts/src/index.ts
@@ -1,1 +1,0 @@
-
